+++ src/components/SearchBar/SearchBar.tsx
@@ -5,7 +5,7 @@ import inputStyle from '~global/scss/Input.module.scss';
 import styles from './SearchBar.module.scss';
 
 import { HomeContext } from '~context/homePageContext';
-import { checkSearch } from '~helpers/checkSearch';
+import { checkSearch } from '~helpers/_index';
 
 export const SearchBar: FC = () => {
   const { cardsState, searchState, isFetchingState } = useContext(HomeContext);
