+++ src/components/_index.ts
@@ -3,3 +3,4 @@ export { Layout } from '~compos/Layout/Layout';
 export { SearchBar } from '~compos/SearchBar/SearchBar';
 export { CardsList } from '~compos/cards/CardsList/CardsList';
 export { Spinner } from '~compos/cards/Spinner/Spinner';
+export { Product } from '~compos/product/Product/Product';
